+++ x11/ui.cpp
@@ -933,7 +933,7 @@ void Ui::menu_enemies_refill_CB(void* value,Viewport*,void* closure) {
   UiP ui = (UiP)closure;
   ui->settingsChanges |= UIenemiesRefill;
   // SGI compiler says you can't cast void* to Boolean.
-  Boolean val = (Boolean)(int)value;
+  Boolean val = (Boolean)(intptr_t)value;
   ui->settings.enemiesRefill = val;
 }
 
@@ -942,7 +942,7 @@ void Ui::menu_enemies_refill_CB(void* value,Viewport*,void* closure) {
 void Ui::menu_controls_CB(void* value,Viewport* vPort,void* closure) {
   UiP ui = (UiP)closure;
   int dpyNum = vPort->get_dpy_num();
-  Boolean val = (Boolean)(int)value;
+  Boolean val = (Boolean)(intptr_t)value;
   if (val) {
     XMapWindow(ui->xvars.dpy[dpyNum],ui->xdata.controls[dpyNum]);
   }
@@ -958,7 +958,7 @@ void Ui::menu_learn_controls_CB(void* value,
                                 Viewport* vPort,void* closure) {
   UiP ui = (UiP)closure;
   int dpyNum = vPort->get_dpy_num();
-  Boolean active = (Boolean)(int)value;
+  Boolean active = (Boolean)(intptr_t)value;
 
   if (active) {
     ui->lControls[dpyNum].input = ui->lControls[dpyNum].key = 
@@ -997,7 +997,7 @@ void Ui::menu_learn_controls_CB(void* value,
 
 void Ui::menu_scenarios_CB(void* value,Viewport*,void* closure) {
   UiP ui = (UiP)closure;
-  Boolean val = (Boolean)(int)value;
+  Boolean val = (Boolean)(intptr_t)value;
   if (val) {
     ui->settingsChanges |= UIstyle;
     ui->settings.style = SCENARIOS;
@@ -1008,7 +1008,7 @@ void Ui::menu_scenarios_CB(void* value,Viewport*,void* closure) {
 
 void Ui::menu_levels_CB(void* value,Viewport*,void* closure) {
   UiP ui = (UiP)closure;
-  Boolean val = (Boolean)(int)value;
+  Boolean val = (Boolean)(intptr_t)value;
   if (val) {
     ui->settingsChanges |= UIstyle;
     ui->settings.style = LEVELS;
@@ -1019,7 +1019,7 @@ void Ui::menu_levels_CB(void* value,Viewport*,void* closure) {
 
 void Ui::menu_kill_CB(void* value,Viewport*,void* closure) {
   UiP ui = (UiP)closure;
-  Boolean val = (Boolean)(int)value;
+  Boolean val = (Boolean)(intptr_t)value;
   if (val) {
     ui->settingsChanges |= UIstyle;
     ui->settings.style = KILL;
@@ -1030,7 +1030,7 @@ void Ui::menu_kill_CB(void* value,Viewport*,void* closure) {
 
 void Ui::menu_duel_CB(void* value,Viewport*,void* closure) {
   UiP ui = (UiP)closure;
-  Boolean val = (Boolean)(int)value;
+  Boolean val = (Boolean)(intptr_t)value;
   if (val) {
     ui->settingsChanges |= UIstyle;
     ui->settings.style = DUEL;
@@ -1041,7 +1041,7 @@ void Ui::menu_duel_CB(void* value,Viewport*,void* closure) {
 
 void Ui::menu_extended_CB(void* value,Viewport*,void* closure) {
   UiP ui = (UiP)closure;
-  Boolean val = (Boolean)(int)value;
+  Boolean val = (Boolean)(intptr_t)value;
   if (val) {
     ui->settingsChanges |= UIstyle;
     ui->settings.style = EXTENDED;
@@ -1052,7 +1052,7 @@ void Ui::menu_extended_CB(void* value,Viewport*,void* closure) {
 
 void Ui::menu_training_CB(void* value,Viewport*,void* closure) {
   UiP ui = (UiP)closure;
-  Boolean val = (Boolean)(int)value;
+  Boolean val = (Boolean)(intptr_t)value;
   if (val) {
     ui->settingsChanges |= UIstyle;
     ui->settings.style = TRAINING;
@@ -1073,7 +1073,7 @@ void Ui::menu_quanta_CB(void* value,Viewport*,void* closure) {
 void Ui::menu_cooperative_CB(void* value,Viewport*,void* closure) {
   UiP ui = (UiP)closure;
   ui->settingsChanges |= UIcooperative;
-  Boolean val = (Boolean)(int)value;
+  Boolean val = (Boolean)(intptr_t)value;
   ui->settings.cooperative = val;
 }
 
@@ -1082,7 +1082,7 @@ void Ui::menu_cooperative_CB(void* value,Viewport*,void* closure) {
 void Ui::menu_help_CB(void* value,Viewport* vPort,void* closure) {
   UiP ui = (UiP)closure;
   int dpyNum = vPort->get_dpy_num();
-  Boolean active = (Boolean)(int)value;
+  Boolean active = (Boolean)(intptr_t)value;
 
   if (active) {
     XMapWindow(ui->xvars.dpy[dpyNum],ui->xdata.help[dpyNum]);
@@ -1097,7 +1097,7 @@ void Ui::menu_help_CB(void* value,Viewport* vPort,void* closure) {
 void Ui::status_weapon_CB(void* value,Viewport* vPort,void*) {
   // Some mouse controls.
   //  UiP ui = (UiP)closure;
-  int button = (int)value;
+  intptr_t button = (intptr_t)value;
 
   switch (button) {
   case Button1:
@@ -1116,7 +1116,7 @@ void Ui::status_weapon_CB(void* value,Viewport* vPort,void*) {
 
 void Ui::status_item_CB(void* value,Viewport* vPort,void*) {
   //  UiP ui = (UiP)closure;
-  int button = (int)value;
+  intptr_t button = (intptr_t)value;
 
   switch (button) {
   case Button1:
